requirements.txt
streamlit
pandas
numpy
altair
joblib
scikit-learn
together
openai
google-api-python-client
beautifulsoup4
requests
python-dotenv
streamlit-mic-recorder
streamlit-TTS
gTTS
pydub
sounddevice
langdetect
watchdog